Move session ownership behind its domain boundary
Constraint: The existing src/session-manager import path has broad fanout and is part of package/runtime compatibility.\nRejected: Rewriting every caller in the same PR | it creates broad noisy churn without adding behavior confidence.\nRejected: Leaving the implementation at src/session-manager.ts | it keeps session lifecycle as a root module after the session folder exists.\nConfidence: high\nScope-risk: moderate\nDirective: Keep src/session-manager.ts as a compatibility shim until callers are migrated in narrower ownership batches.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/tools/manifest-only-registration.test.ts tests/session/manager-contracts.test.ts tests/session/manager-tenant.test.ts tests/session/manager-target-creation.test.ts tests/session/manager-evict-target.test.ts tests/storage-state/session-manager-wiring.test.ts tests/cli/entrypoint-parity.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: git diff --check\nNot-tested: Full Jest suite.
